Add tapipy as a package and refactor
